feat(contract): consolidate DeclineReason into @gamedevpl/contract
Phase 1, eleventh slice. The six reasons a reviewer can decline a proposal
(not_the_direction, duplicate, quality, off_topic, unsafe, infringing)
were defined once as a const array in apps/api/src/proposal-state.ts, then
hand-typed identically in apps/web/src/proposalsApi.ts, then hand-listed a
third time as a runtime array in apps/web/src/ProposalReviewCard.tsx for
its decline-reason <select> — three copies of the same closed vocabulary,
none referencing either of the others.

Added packages/contract/src/decline-reason.ts with DECLINE_REASONS/
DeclineReason, following the same const-array-with-derived-type shape as
proposal-state.ts's own PROPOSAL_STATES, plus a matching test file.
proposal-state.ts now re-exports the contract's values instead of
declaring them (proposals.ts, store.ts, proposal-routes.ts, and its own
tests all import DeclineReason/DECLINE_REASONS from ./proposal-state.js
already and need no changes). proposalsApi.ts imports and re-exports the
type from @gamedevpl/contract, matching its existing pattern for other
contract types; ProposalReviewCard.tsx drops its hand-listed array
entirely and imports DECLINE_REASONS from ./proposalsApi.js instead,
matching how it already imports everything else proposal-related from
that file rather than from the contract directly.
