+++ eslint-rules/module-size-baseline.json
@@ -305,7 +305,7 @@
     "apps/api/src/proposal-routes.test.ts": 378,
     "apps/api/src/proposal-routes.ts": 457,
     "apps/api/src/proposal-state.test.ts": 181,
-    "apps/api/src/proposal-state.ts": 286,
+    "apps/api/src/proposal-state.ts": 269,
     "apps/api/src/proposals.test.ts": 528,
     "apps/api/src/proposals.ts": 713,
     "apps/api/src/published-slugs.test.ts": 175,
@@ -689,7 +689,7 @@
     "apps/web/src/priorRoundDismiss.ts": 28,
     "apps/web/src/ProposalDiffView.tsx": 86,
     "apps/web/src/ProposalReviewCard.test.tsx": 117,
-    "apps/web/src/ProposalReviewCard.tsx": 205,
+    "apps/web/src/ProposalReviewCard.tsx": 197,
     "apps/web/src/ProposalReviewPanel.tsx": 56,
     "apps/web/src/proposalsApi.ts": 262,
     "apps/web/src/ProposalsPage.tsx": 140,
@@ -873,11 +873,13 @@
     "apps/world/src/snapshot-store.ts": 92,
     "packages/contract/src/agent-channel-routes.test.ts": 20,
     "packages/contract/src/agent-channel-routes.ts": 39,
+    "packages/contract/src/decline-reason.test.ts": 8,
+    "packages/contract/src/decline-reason.ts": 10,
     "packages/contract/src/gate-progress.test.ts": 8,
     "packages/contract/src/gate-progress.ts": 3,
     "packages/contract/src/gate-status.test.ts": 49,
     "packages/contract/src/gate-status.ts": 25,
-    "packages/contract/src/index.ts": 8,
+    "packages/contract/src/index.ts": 9,
     "packages/game-generator/src/engine/engine.test.ts": 35,
     "packages/game-generator/src/index.ts": 2,
     "packages/game-generator/src/mock.test.ts": 67,
